requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
websockets==12.0
eth-account==0.10.0
web3==6.16.0
pydantic==2.5.3
pydantic-settings==2.1.0
